refactor: changes site.metadata to constans file
